migration fix for codeship (dry run)

+++ smallslive/artists/migrations/0011_auto__del_artisttype__add_instrument.py
@@ -8,8 +8,9 @@ from django.db import models
 class Migration(SchemaMigration):
 
     def forwards(self, orm):
-        old_content_type = orm['contenttypes.ContentType'].objects.get(app_label='artists', model='artisttype')
-        old_content_type.delete()
+        if not db.dry_run:
+            old_content_type = orm['contenttypes.ContentType'].objects.get(app_label='artists', model='artisttype')
+            old_content_type.delete()
 
         # Renaming ArtistType to Instrument
         db.rename_table(u'artists_artisttype', u'artists_instrument')
@@ -37,7 +38,8 @@ class Migration(SchemaMigration):
         db.rename_table(u'artists_instrument', u'artists_artisttype')
         db.send_create_signal(u'artists', ['ArtistType'])
 
-        orm['contenttypes.ContentType'].objects.create(app_label='artists', model='artisttype')
+        if not db.dry_run:
+            orm['contenttypes.ContentType'].objects.create(app_label='artists', model='artisttype')
